+++ CHANGELOG.md
@@ -30,10 +30,12 @@ VDisclosureGroup
 VModal
 
 - `contentMargins` are changed from `15`s to `zero` in `VModalUIModel`. This configuration supports list with already-padded rows. But in case of non-list content, additional padding must be used.
+- `VModalUIModel.noHeaderLabel` is added
 
 VBottomSheet
 
 - `contentMargins` are changed from `15`s to `zero` in `VBottomSheetUIModel`. This configuration supports list with already-padded rows. But in case of non-list content, additional padding must be used.
+- `VBottomSheetUIModel.scrollableContent`, `VBottomSheetUIModel.noHeaderLabel`, and `VBottomSheetUIModel.scrollableContentNoHeaderLabel` are added
 
 VSideBar